+++ src/Http/OAuth2/Providers/LeagueProvider.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Larabros\Elogram\Http\OAuth2\Providers;
+
+use League\OAuth2\Client\Provider\Instagram;
+
+/**
+ * An OAuth2 provider for the League Instagram OAuth2 package.
+ *
+ * @package    Elogram
+ * @author     Hassan Khan <contact@hassankhan.me>
+ * @link       https://github.com/larabros/elogram
+ * @license    MIT
+ */
+class LeagueAdapter extends Instagram implements AdapterInterface
+{
+}